Configure NVIM useable system
Major inspiration from: https://dev.to/slydragonn/how-to-set-up-neovim-for-windows-and-linux-with-lua-and-packer-2391

Add an lsp
Add a status bar
Custom function and mappings for turning certaing stuff on and off.
All edits are under neovim configs. Added a lot of configs under
nvim/lua and just nvim/init.vim is for adding more files to read.

Using nvim/lua/init.lua for some initial configs via lua instead of
init.vim
Using nvim/lua/custom-functions.lua as a place for personal functions from
now on. Have one for toggling status line.
Using nvim/lua/plugins.lua for neovim plugins via Packer. This is not a
Using nvim/lua/maps.lua for key mapping. This means there is now two
placs for key maps for nvim, init.nvim and maps.lua. This is pretty
terrible and I recognize this, but will leave it for now.

Most configurations were copied from the site mentioned above.